fix: bring back disabling the submit button
My theory on what's happening with the double-creates is that she clicks the button, but since she leaves the page open all the time (right? that was her workflow if I remember right?) sometimes (or most times) App Engine has to start the app up from scratch which takes about 5 seconds. In those five seconds she gets antsy and then clicks the button again, but the first request has seemingly made it through so then it goes through again and the two POs end up getting the same "next po number" and create at the same time with all the same fields.

Now, we disable the create button forever and change the text to "Creating..."; this will clear when the form either goes through successfully or the page reloads with an error.